Add ForumUser service, update UserRegdate type
- Changed `UserRegdate` column type in `20241029184053_AddForumTables.cs` migration file from `int` to `long`.
- Updated `UserRegdate` property in `ForumUser` class from `int` to `long`.
- Added and registered `IForumUserService` in `Program.cs`.
- Modified `ForumConfigService` constructor to accept `DigiozForumContext` parameter.
- Added `UserHelper` class in `digioz.Forum.Helpers` namespace.
- Added `ForumUserService` class in `digioz.Forum.Services` namespace, implementing `IForumUserService`.
- Added `IForumUserService` interface in `digioz.Forum.Services.Interfaces` namespace.

+++ source/digioz.Forum/digioz.Forum/Data/Migrations/20241029184053_AddForumTables.cs
@@ -1199,7 +1199,7 @@ namespace digioz.Forum.Data.Migrations
                     UserPermissions = table.Column<string>(type: "text", nullable: false),
                     UserPermFrom = table.Column<long>(type: "bigint", nullable: false, defaultValue: 0),
                     UserIp = table.Column<string>(type: "varchar(40)", nullable: false, defaultValue: ""),
-                    UserRegdate = table.Column<int>(type: "int", nullable: false, defaultValue: 0),
+                    UserRegdate = table.Column<long>(type: "bigint", nullable: false, defaultValue: 0),
                     UserName = table.Column<string>(type: "varchar(255)", nullable: false, defaultValue: ""),
                     UserNameClean = table.Column<string>(type: "varchar(255)", nullable: false, defaultValue: ""),
                     UserEmail = table.Column<string>(type: "varchar(100)", nullable: false, defaultValue: ""),
